Removed 'force_rate' as it didn't make sense in our case.
